refactor: move llama-server-context out of controller
